sleepyEyes og energyState
i denne ændring har jeg tilføjet energyState der ville bestemme om vores kæledyr sover eller er vågen. det er den samme værdi som vi bruger til at definere om øjnene skal lukke når den sover og om de er åbne når den er vågen.

+++ DigitalPet_EigilMads/Aktivities.pde
@@ -1,17 +1,17 @@
 class Activity {
   String title;
-  float energyChange;
-  
+  int energyChange;
 
 
-  Activity(String title, float energyChange) {
+
+  Activity(String title, int energyChange) {
     this.title = title;
     this.energyChange = energyChange;
   }
 
   void sleep(DigitalPet pet) {
     pet.changeEnergy(energyChange);
-
+   
   }
